fix(typesystem): the two bounds 194a claimed now hold (its security pass) — PatternGuard.matches catches the StackOverflowError an alternation inside a repetition ((a|b)*) raises thousands of reads before the budget and reports it as the budget refusal, so no Error reaches a REST 500, the MCP dispatcher's Exception catch or a schedule run retried forever (red: StackOverflowError escaped on a 512 KiB stack); the precision check's precision() − scale() is done in Long — with scale() = −2147483647 it overflowed Int and accepted exactly the values with the most digits (red: 1e2147483647 accepted by BIGDECIMAL(10,2)); plus two observations: \p{L}+ is no longer refused as a possessive quantifier (the scan reads the property's braces as the parser does), and a bound renders through BigDecimal.toString() in messages — 1e2147483647 as a min was an OutOfMemoryError that killed the test JVM (red: the executor died). typesystem 1192/0, pipeline-contract 484/0, ktlint + detekt executed (#194)

+++ modules/typesystem/src/main/kotlin/co/datapipelines/typesystem/ParameterValueValidator.kt
@@ -134,8 +134,10 @@ class ParameterValueValidator(
             return violation("scale", "$shape takes at most $scale decimal place(s); got $places — nothing is rounded")
         }
         val precision = declaration.precision ?: return null
-        val integerDigits = if (stripped.signum() == 0) 0 else maxOf(0, stripped.precision() - stripped.scale())
-        if (integerDigits > precision - scale) {
+        // In Long: scale() can be -2147483647, and precision() - scale() in Int overflowed to accept
+        // exactly the values with the most digits (the 194a security pass, finding 2).
+        val integerDigits = if (stripped.signum() == 0) 0L else maxOf(0L, stripped.precision().toLong() - stripped.scale().toLong())
+        if (integerDigits > (precision - scale).toLong()) {
             return violation("precision", "$shape holds at most ${precision - scale} integer digit(s); got $integerDigits")
         }
         return null
